Refactoring in an attempt to get rid of graphics dependencies and Orbiter.h in Psys; Pulled TimeData into its own file so it can be included without pulling the whole Orbiter.h; Removed some unused members/parameters;
